chore(text-input): move to packages
