+++ packages/components/inputs/text-input/src/index.js
@@ -1,0 +1,3 @@
+export { default } from './text-input';
+
+export { default as version } from './version';
