+++ packages/components/inputs/text-input/src/text-input.js
@@ -3,7 +3,7 @@ import PropTypes from 'prop-types';
 import requiredIf from 'react-required-if';
 import { filterDataAttributes } from '@commercetools-uikit/utils';
 import Constraints from '@commercetools-uikit/constraints';
-import { getInputStyles } from '../styles';
+import { getInputStyles } from '../../../../../src/components/inputs/styles';
 
 const TextInput = (props) => {
   return (
