--- src/components/inputs/text-input/index.js
@@ -1,1 +1,0 @@
-export { default } from './text-input';
